Adding some screenshots.

+++ TClient/src/Tester.java
@@ -18,9 +18,17 @@ public class Tester {
 	
 		TClient client = new TClient();
 		
-		Files.deleteIfExists(Paths.get("./Output/Debian-ISO.iso"));
-		client.downloadFile(Paths.get("./Debian ISO/debian-10.4.0-amd64-netinst.iso.torrent"), Paths.get("./Output/Debian-ISO.iso"), PieceSelectionPolicy.RarestFirst);
 		
+		if(args.length < 1) {
+			System.out.println("Error: No torrent supplied!");
+		} else {
+			// Remove previous output
+			Files.deleteIfExists(Paths.get("./Output/Debian-ISO.iso"));
+			
+			client.downloadFile(Paths.get(args[0]), Paths.get("../Output/Debian-ISO.iso"), PieceSelectionPolicy.RarestFirst);
+			//client.downloadFile(Paths.get("./Debian ISO/debian-10.4.0-amd64-netinst.iso.torrent"), Paths.get("./Output/Debian-ISO.iso"), PieceSelectionPolicy.RarestFirst);
+
+		}
 		
 	}
